BookAnAppointment_react/backend/requirements.txt
celery
Django
django-cors-headers
django-filter
djangorestframework
eventlet
Pillow
redis
python-decouple
